+++ src/components/toolbar/tools/HeadingTool.js
@@ -90,6 +90,7 @@ class HeadingTool extends MakeTool {
 
         this.editor.scrollToView();
         this.editor.render();
+        this.editor.triggerChange();
     }
 
     applySyntax() {}